Fix choropleth64 layer (#172)
Make the attribute.value array matching the data type specified with attributeManager

+++ src/layers/fp64/choropleth-layer/choropleth-layer.js
@@ -202,7 +202,7 @@ export default class ChoroplethLayer64 extends Layer {
       }
     );
 
-    attribute.value = new Float32Array(flattenDeep(colors));
+    attribute.value = new Uint8Array(flattenDeep(colors));
   }
 
   // Override the default picking colors calculation
@@ -220,7 +220,7 @@ export default class ChoroplethLayer64 extends Layer {
       }
     );
 
-    attribute.value = new Float32Array(flattenDeep(colors));
+    attribute.value = new Uint8Array(flattenDeep(colors));
   }
 }
